tests/wpt/web-platform-tests/html/dom/elements/elements-in-the-dom/historical.html

<!doctype html>
<title>Historical HTMLElement features</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<body>
<script>
[
  // https://github.com/whatwg/html/commit/389ec2620d89e9480ef8847bf016abdfa92427bc
  "commandType",
  "commandLabel",
  "commandIcon",
  "commandHidden",
  "commandDisabled",
  "commandChecked",
  "commandTriggers",
  // https://github.com/whatwg/html/commit/5ddfc78b1f82e86cc202d72ccc752a0e15f1e4ad
  "inert",
].forEach(function(member) {
  test(function() {
    assert_false(member in document.body);
    assert_false(member in document.createElement('div'));
  }, 'HTMLElement member must be nuked: ' + member);
});
</script>
